# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/blocks": {
            "post": {
                "description": "Create a new block with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "Create a new block",
                "parameters": [
                    {
                        "description": "Block Data",
                        "name": "block",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PostBlockRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/blocks/{hash}": {
            "get": {
                "description": "Get a block by its hash",
                "tags": [
                    "blocks"
                ],
                "summary": "Get a block by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Block Hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Block": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.PostBlockRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "block": {
                    "$ref": "#/definitions/main.Block"
                },
                "hash": {
                    "type": "string"
                }
            }
        }
    }
}